Fixed #3304 - registration of rider consumers, sagas, etc. were combined with bus registrations.
